+++ android/src/main/java/com/reactlibrary/MatrixSdkModule.java
@@ -1,6 +1,10 @@
 package com.reactlibrary;
 
+import android.content.Context;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
 import android.net.Uri;
+import android.os.Build;
 import android.util.Log;
 
 import androidx.annotation.Nullable;
@@ -22,6 +26,7 @@ import org.json.JSONException;
 import org.matrix.androidsdk.HomeServerConnectionConfig;
 import org.matrix.androidsdk.MXDataHandler;
 import org.matrix.androidsdk.MXSession;
+import org.matrix.androidsdk.data.Pusher;
 import org.matrix.androidsdk.data.Room;
 import org.matrix.androidsdk.data.RoomState;
 import org.matrix.androidsdk.data.store.MXFileStore;
@@ -31,6 +36,7 @@ import org.matrix.androidsdk.rest.model.CreateRoomParams;
 import org.matrix.androidsdk.rest.model.CreateRoomResponse;
 import org.matrix.androidsdk.rest.model.CreatedEvent;
 import org.matrix.androidsdk.rest.model.Event;
+import org.matrix.androidsdk.rest.model.PushersResponse;
 import org.matrix.androidsdk.rest.model.TokensChunkEvents;
 import org.matrix.androidsdk.rest.model.User;
 import org.matrix.androidsdk.rest.model.login.Credentials;
@@ -403,6 +409,81 @@ public class MatrixSdkModule extends ReactContextBaseJavaModule implements Lifec
                 });
     }
 
+    @ReactMethod
+    public void registerPushNotifications(String displayName, String pushServiceUrl, String token, Promise promise) {
+        if (mxSession == null) {
+            promise.reject(E_MATRIX_ERROR, "client is not connected yet");
+            return;
+        }
+
+        // First check whether a pusher is already registered
+        mxSession.getPushersRestClient()
+                .getPushers(new RejectingOnErrorApiCallback<PushersResponse>(promise) {
+                    @Override
+                    public void onSuccess(PushersResponse info) {
+                        List<Pusher> pushers = info.pushers;
+                        if(pushers.isEmpty()) {
+                            // register push notifications
+                            addHttpPusher(displayName, pushServiceUrl, token, promise);
+                        } else {
+                            // check pushers:
+                            boolean needToRegister = true;
+                            for(Pusher pusher : pushers) {
+                                Log.d(TAG, pusher.toString());
+                                if(pusher.pushkey.equals(token)) {
+                                    needToRegister = false;
+                                    break;
+                                }
+                            }
+                            if(needToRegister) {
+                                addHttpPusher(displayName, pushServiceUrl, token, promise);
+                            } else {
+                                // pusher already setup :)
+                                promise.resolve(null);
+                            }
+                        }
+                    }
+                });
+    }
+
+    /**
+     * internal, assumes that you have checked that session is active.
+     * Implementation from {https://github.com/vector-im/riot-android/blob/develop/vector/src/main/java/im/vector/push/PushManager.java}
+     */
+    private void addHttpPusher(String displayName, String pushServiceUrl, String token, Promise promise) {
+        Context mContext = reactContext.getApplicationContext();
+        PackageInfo pInfo = null;
+        try {
+            pInfo = mContext.getPackageManager().getPackageInfo(mContext.getPackageName(), 0);
+        } catch (PackageManager.NameNotFoundException e) {
+            e.printStackTrace();
+            promise.reject(e);
+            return;
+        }
+        String mPusherAppName = pInfo.packageName;
+        String mPusherLang = mContext.getResources().getConfiguration().locale.getLanguage();
+        String mBasePusherDeviceName = "Cuvent_" +Build.MODEL.trim();
+
+        mxSession.getPushersRestClient()
+                .addHttpPusher(
+                        token,
+                        mPusherAppName,
+                        computePushTag(mxSession),
+                        mPusherLang,
+                        displayName,
+                        mBasePusherDeviceName,
+                        pushServiceUrl,
+                        false,
+                        false,
+                        new RejectingOnErrorApiCallback<Void>(promise) {
+                            @Override
+                            public void onSuccess(Void info) {
+                                promise.resolve(info);
+                            }
+                        }
+                );
+    }
+
     @Override
     public void onHostResume() {
 
@@ -425,5 +506,23 @@ public class MatrixSdkModule extends ReactContextBaseJavaModule implements Lifec
         reactContext
                 .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
                 .emit(eventName, params);
+    }
+
+
+    /**
+     * Compute the profileTag for a session
+     *
+     * @param session the session
+     * @return the profile tag
+     */
+    private static String computePushTag(final MXSession session) {
+        String tag = "cuvent_mobile_" + Math.abs(session.getMyUserId().hashCode());
+
+        // tag max length : 32 bytes
+        if (tag.length() > 32) {
+            tag = Math.abs(tag.hashCode()) + "";
+        }
+
+        return tag;
     }
 }