+++ ios/MatrixSdk.m
@@ -54,4 +54,6 @@ RCT_EXTERN_METHOD(sendMessageToRoom:(NSString *)roomId messageType:(NSString *)m
 
 RCT_EXTERN_METHOD(sendReadReceipt:(NSString *)roomId eventId:(NSString *)eventId resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
 
+RCT_EXTERN_METHOD(registerPushNotifications:(NSString *)displayName pushServiceUrl:(NSString *)pushServiceUrl token:(NSString *)token resolver:(RCTPromiseResolveBlock)resolve rejecter:(RCTPromiseRejectBlock)reject)
+
 @end